+++ pyexasol_utils/bench_sql/t4_sort.sql
@@ -1,8 +1,8 @@
-SELECT low_cardinality_col, rn
+SELECT low_cardinality_col, id
 FROM (
     SELECT low_cardinality_col
-        , rn
-        , ROW_NUMBER() OVER (PARTITION BY low_cardinality_col ORDER BY rn ASC) AS rnum
+        , id
+        , ROW_NUMBER() OVER (PARTITION BY low_cardinality_col ORDER BY id ASC) AS rnum
     FROM test_data_1
 )
 WHERE rnum=1